Updated impulse data to N instead of impulse length

+++ Set3/fft_convolve_cuda.cu
@@ -61,10 +61,10 @@ cudaProdScaleKernel(const cufftComplex *raw_data, const cufftComplex *impulse_v,
     while(index < padded_length)
     {
         out_data[index].x = (raw_data[index].x * impulse_v[index].x -
-            raw_data[index].y * impulse_v[index].y) / (float) padded_length;
+            raw_data[index].y * impulse_v[index].y) / padded_length;
 
         out_data[index].y = (raw_data[index].x * impulse_v[index].y +
-            raw_data[index].y * impulse_v[index].x) / (float) padded_length;
+            raw_data[index].y * impulse_v[index].x) / padded_length;
 
         // Compute next index for arbitrary amount of threads
         index += blockDim.x * gridDim.x;
